feat: move PartialPacket in protocol
